refactor: extract entry actions `favorite` and `delete` into components
